+++ CHANGELOG.md
@@ -12,6 +12,8 @@ with respect to its command line interface and HTTP interface
   was sometimes inaccurately recorded. Real world implications of this are not
   completely clear, users are not expected to notice much difference.
 * Client: sous metadata set would panic if no existing metadata was present in manifest prior to set
+* Client: sous jenkins cli to use manifest to generate JenkinsPipeline file
+
 ### Changed
 * Client: `SOUS_USE_SOUS_SEVER` env var must now be exactly uppercase `YES`
   in order to be considered "on". Previously any value, even empty string
@@ -26,6 +28,7 @@ with respect to its command line interface and HTTP interface
 ### Changed
 * Tests: Travis build times reduced by running sets of tests in parallel.
   This allows us to produce relese binaries, see note below in 0.5.116.
+
 
 ## [0.5.116](//github.com/opentable/sous/compare/0.5.115...0.5.116)
 ### Changed